requirements.txt
Django==4.2.15
djangorestframework==3.15.2
python-dotenv==1.0.1
Pillow==10.4.0